Give ol.style.Atlas its own file
